+++ src/sampling/entropic_sampling.rs
@@ -1,7 +1,0 @@
-mod entropic_adaptive;
-mod entropic;
-mod traits;
-
-pub use entropic::*;
-pub use entropic_adaptive::*;
-pub use traits::*;