+++ src/sampling/glue.rs
@@ -1,7 +1,0 @@
-mod glue_wl;
-pub(crate) mod glue_helper;
-mod glue_entropic;
-
-pub use glue_wl::*;
-pub use glue_entropic::*;
-pub use glue_helper::GlueErrors;