+++ src/sampling/heatmap.rs
@@ -1,9 +1,0 @@
-mod heatmap;
-mod helper;
-mod gnuplot;
-mod heatmap_float;
-
-pub use heatmap::*;
-pub use helper::*;
-pub use gnuplot::*;
-pub use heatmap_float::*;