+++ src/sampling/wang_landau.rs
@@ -1,9 +1,0 @@
-mod wang_landau_adaptive;
-mod helper;
-mod wang_landau;
-mod traits;
-
-pub use wang_landau_adaptive::*;
-pub use helper::*;
-pub use wang_landau::*;
-pub use traits::*;